Added check for yazi to be installed in yy.fish

+++ files/private_dot_config/private_fish/functions/yy.fish
@@ -1,9 +1,13 @@
 function yy
-set tmp (mktemp -t "yazi-cwd.XXXXXX")
-yazi $argv --cwd-file="$tmp"
-if set cwd (command cat -- "$tmp"); and [ -n "$cwd" ]; and [ "$cwd" != "$PWD" ]
-builtin cd -- "$cwd"
-end
-rm -f -- "$tmp"
+  if not command -q yazi
+    echo 'Error: yazi not installed'
+    exit 1
+  end
+  set tmp (mktemp -t "yazi-cwd.XXXXXX")
+  yazi $argv --cwd-file="$tmp"
+  if set cwd (command cat -- "$tmp"); and [ -n "$cwd" ]; and [ "$cwd" != "$PWD" ]
+    builtin cd -- "$cwd"
+  end
+  rm -f -- "$tmp"
 end
